requirements.txt
pyrogram
tgcrypto
requests
python-dotenv
google-generativeai
ipython
pillow
